# back/src/modules/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Swagger for Cocktail API",
    "description": "Open API docs with Swagger",
    "license": { "name": "MIT", "url": "http://localhost:5000" }
  },
  "host": "localhost:5000",
  "basePath": "/",
  "tags": [
    { "name": "User", "description": "User 로그인, 회원가입 API" },
    { "name": "Cocktail", "description": "칵테일 조회 API" },
    { "name": "Like", "description": "유저 칵테일 LIKE API" }
  ],
  "schemes": ["HTTP"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/": { "get": { "tags": ["Main"], "summary": "Main Page Get", "responses": { "200": { "description": "OK" } } } },
    "/cocktail": {
      "post": {
        "tags": ["Cocktail"],
        "summary": "cocktail 조회",
        "parameters": [{ "in": "body", "name": "body", "schema": { "$ref": "#/definitions/Cocktail" } }],
        "responses": { "201": { "description": "OK" }, "404": { "description": "Failed. Bad Request" } }
      }
    },
    "/populateCocktail": {
      "post": {
        "tags": ["Cocktail"],
        "summary": "cocktail 조회 + populate",
        "parameters": [{ "in": "body", "name": "body", "schema": { "$ref": "#/definitions/Cocktail" } }],
        "responses": { "201": { "description": "OK" }, "404": { "description": "Failed. Bad Request" } }
      }
    },
    "/cocktaillist": {
      "get": {
        "tags": ["Cocktail"],
        "summary": "cocktail 조회",
        "responses": { "201": { "description": "OK" }, "404": { "description": "Failed. Bad Request" } }
      }
    },
    "/addLike/{id}": {
      "post": {
        "tags": ["Like"],
        "summary": "좋아요 추가",
        "parameters": [
          { "in": "path", "name": "id" },
          {
            "name": "Authorization",
            "in": "header",
            "description": "JWT access token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": { "201": { "description": "OK" }, "404": { "description": "Failed. Bad Request" } }
      }
    },
    "/deleteLike/{id}": {
      "delete": {
        "tags": ["Like"],
        "summary": "좋아요 삭제",
        "parameters": [
          { "in": "path", "name": "id" },
          {
            "name": "Authorization",
            "in": "header",
            "description": "JWT access token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": { "201": { "description": "OK" }, "404": { "description": "Failed. Bad Request" } }
      }
    },
    "/likeliet": {
      "get": {
        "tags": ["Like"],
        "summary": "좋아요 전체 리스트",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "JWT access token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": { "201": { "description": "OK" }, "404": { "description": "Failed. Bad Request" } }
      }
    },
    "/cocktailLike/{id}": {
      "get": {
        "tags": ["Like"],
        "summary": "칵테일이 받은 좋아요 리스트",
        "parameters": [
          { "in": "path", "name": "id" },
          {
            "name": "Authorization",
            "in": "header",
            "description": "JWT access token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": { "201": { "description": "OK" }, "404": { "description": "Failed. Bad Request" } }
      }
    },
    "/userLike/{id}": {
      "get": {
        "tags": ["Like"],
        "summary": "유저가 좋아요한 리스트 목록",
        "parameters": [
          { "in": "path", "name": "id" },
          {
            "name": "Authorization",
            "in": "header",
            "description": "JWT access token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": { "201": { "description": "OK" }, "404": { "description": "Failed. Bad Request" } }
      }
    },
    "/rank10": {
      "get": {
        "tags": ["Rank"],
        "summary": "cocktail 조회",
        "responses": { "201": { "description": "OK" }, "404": { "description": "Failed. Bad Request" } }
      }
    },
    "/register": {
      "post": {
        "tags": ["User"],
        "summary": "register user Post",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "email": "body",
            "password": "body",
            "schema": { "$ref": "#/definitions/User" }
          }
        ],
        "responses": { "201": { "description": "OK" }, "404": { "description": "Failed. Bad Request" } }
      }
    },
    "/login": {
      "post": {
        "tags": ["User"],
        "summary": "login user Post",
        "parameters": [
          {
            "in": "body",
            "email": "body",
            "password": "body",
            "name": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/User" }
          }
        ],
        "responses": { "201": { "description": "OK" }, "404": { "description": "Failed. Bad Request" } }
      }
    }
  },
  "definitions": {
    "Cocktail": {
      "properties": {
        "name": { "type": "string" },
        "ingredient": { "type": "string" },
        "imageUrl": { "type": "string" },
        "rank": { "type": "string" },
        "taste": { "type": "string" },
        "descriptoin": { "type": "string" },
        "userId": { "type": "string" }
      }
    },
    "Like": {
      "properties": {
        "giveUserId": { "type": "string" },
        "getCocktailId": { "type": "string" },
        "name": { "type": "string" }
      }
    },
    "Rank": {
      "properties": {
        "name": { "type": "string" },
        "visitors": { "type": "string" },
        "years": { "type": "string" },
        "rank": { "type": "string" }
      }
    },
    "User": {
      "properties": { "email": { "type": "string" }, "password": { "type": "string" }, "name": { "type": "string" } }
    }
  }
}